book_explorer_backend/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Jsonpify==1.5.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
wsproto==1.2.0
